feat: add UIViewController extension
