+++ .gitignore
@@ -1,0 +1,1 @@
+BlueToothScanner/BlueToothScanner.xcodeproj/xcuserdata/tuannv19.xcuserdatad/xcdebugger
