+++ BlueToothScanner/BlueToothScanner/Viewcontrollers/BluetoothViewController/BluetoothViewController.swift
@@ -34,22 +34,14 @@ class BluetoothViewController: UIViewController {
         )
         self.tableView.reloadData()
     }
-    
-    func showAlert(title: String, message: String)  {
-        let alert = UIAlertController(title: title,
-                                      message: message,
-                                      preferredStyle: UIAlertController.Style.alert)
-        alert.addAction(UIAlertAction(title: "Ok",
-                                      style: UIAlertAction.Style.default,
-                                      handler: nil))
-        self.present(alert, animated: true, completion: nil)
-    }
 }
+
 extension BluetoothViewController {
     @IBAction func sliderStartsScanValueChanged(_ sender: UISwitch) {
         self.viewmodel.isScanning = sender.isOn
     }
 }
+
 extension BluetoothViewController: UITableViewDelegate{
     
 }
